Added a record to store margin information and update the temporary calculations to show that.

+++ src/main/java/dev/eamell/terms/cardgenerator/model/Size.java
@@ -1,4 +1,4 @@
 package dev.eamell.terms.cardgenerator.model;
 
-public record Size(int width, int height) {
+public record Size(float width, float height) {
 }
